requirements.txt
python_telegram_bot==13.7
requests==2.26.0